feat: CRM módulo + auth + CI Supabase + refinamentos catálogo
- src/app/crm/: módulo CRM autenticado (layout, dashboard, leads, novo lead)
- src/app/login/, src/app/logout/: fluxo de auth com Supabase
- src/lib/supabase/middleware.ts: protege /crm e redireciona p/ /login
- .github/workflows/supabase-migrate.yml: deploy de migrações via CLI
  (push em main com mudanças em supabase/migrations/** dispara dev;
   workflow_dispatch com choice dev/prod para aplicar manualmente)
- .dockerignore: enxuga imagem de produção (resolve achado #10 da revisão)
- supabase/migrations/0001: adiciona citext (faltava p/ leads.email)
- supabase/migrations/0004: corrige null sem cast em service_addons + dev-web
- supabase/migrations/0006: preços Desenvolvimento Web + 3 pacotes
  Campanha de Lançamento + 3 bundles Branding Project
- supabase/migrations/0007: 3 pacotes RP & Marketing Direto
- supabase/migrations/0008: descrições enriquecidas dos serviços de Agência
- supabase/migrations/0009: índices em FKs + reescreve políticas RLS p/
  usar (select auth.uid()) — corrige unindexed_foreign_keys + auth_rls_initplan

Migrações 0001-0009 já aplicadas em ambos projetos Supabase
(house-mazzutti-dev e house-mazzutti-prod, sa-east-1).

+++ supabase/migrations/0004_seed_catalog.sql
@@ -126,13 +126,13 @@ on conflict do nothing;
 
 -- Add-ons globais (entrega rápida, periculosidade, etc.)
 insert into public.service_addons (service_id, slug, name, description, price_brl, position) values
-  (null, 'entrega-rapida-6h',  'Entrega rápida (até 6h)',  'Entrega das fotos editadas em até 6 horas após o shooting.', null, 10),
-  (null, 'entrega-rapida-24h', 'Entrega rápida (até 24h)', 'Entrega das fotos editadas em até 24 horas após o shooting.', null, 20),
-  (null, 'publicacao-blog',    'Publicação no Blog House', 'Inclusão do material no blog House Mazzutti.',                null, 30),
-  (null, 'publicacao-externa', 'Publicação externa',       'Direitos de publicação em mídia externa.',                    null, 40),
-  (null, 'alimentacao',        'Alimentação',              'Alimentação da equipe.',                                     null, 50),
-  (null, 'risco-periculosidade','Risco / Periculosidade',  'Adicional para locações com risco operacional.',              null, 60),
-  (null, 'equipamento-extra',  'Equipamento extra',        'Equipamento técnico além do padrão.',                         null, 70)
+  (null, 'entrega-rapida-6h',  'Entrega rápida (até 6h)',  'Entrega das fotos editadas em até 6 horas após o shooting.', null::numeric, 10),
+  (null, 'entrega-rapida-24h', 'Entrega rápida (até 24h)', 'Entrega das fotos editadas em até 24 horas após o shooting.', null::numeric, 20),
+  (null, 'publicacao-blog',    'Publicação no Blog House', 'Inclusão do material no blog House Mazzutti.',                null::numeric, 30),
+  (null, 'publicacao-externa', 'Publicação externa',       'Direitos de publicação em mídia externa.',                    null::numeric, 40),
+  (null, 'alimentacao',        'Alimentação',              'Alimentação da equipe.',                                     null::numeric, 50),
+  (null, 'risco-periculosidade','Risco / Periculosidade',  'Adicional para locações com risco operacional.',              null::numeric, 60),
+  (null, 'equipamento-extra',  'Equipamento extra',        'Equipamento técnico além do padrão.',                         null::numeric, 70)
 on conflict do nothing;
 
 ------------------------------------------------------------
@@ -233,9 +233,9 @@ on conflict do nothing;
 insert into public.service_packages (service_id, slug, name, price_brl, includes, position)
 select s.id, p.slug, p.name, p.price_brl, p.includes::jsonb, p.position
 from public.services s, (values
-  ('essential',     'Essential',     null, '["Site institucional / landing","Performance e responsividade","SEO técnico inicial","Integração WhatsApp e Instagram","Formulário de contato"]', 10),
-  ('profissional',  'Profissional',  null, '["Site multi-página com blog","Galerias visuais","Áreas com CTA de conversão","Formulários inteligentes","Integrações de agenda/CRM","Treinamento básico"]', 20),
-  ('premium',       'Premium',       null, '["Plataforma personalizada (WooCommerce ou stack custom)","Design exclusivo","E-commerce / catálogo / agenda","SEO avançado","Integrações ilimitadas (ERPs, marketplaces)","Manutenção mensal opcional"]', 30)
+  ('essential',     'Essential',     null::numeric, '["Site institucional / landing","Performance e responsividade","SEO técnico inicial","Integração WhatsApp e Instagram","Formulário de contato"]', 10),
+  ('profissional',  'Profissional',  null::numeric, '["Site multi-página com blog","Galerias visuais","Áreas com CTA de conversão","Formulários inteligentes","Integrações de agenda/CRM","Treinamento básico"]', 20),
+  ('premium',       'Premium',       null::numeric, '["Plataforma personalizada (WooCommerce ou stack custom)","Design exclusivo","E-commerce / catálogo / agenda","SEO avançado","Integrações ilimitadas (ERPs, marketplaces)","Manutenção mensal opcional"]', 30)
 ) as p(slug, name, price_brl, includes, position)
 where s.slug = 'desenvolvimento-web'
 on conflict (service_id, slug) do nothing;
